managedendpoint: name the error vocabulary, usage, and request parsing
route_error.go holds the OpenAI envelope and every fixed cause; call sites
use names instead of status/code/message triples. usage.go gathers the Usage
type, extraction, validity and pricing (pricing.go folded in). request.go is
body and model parsing. Long compound conditions become named predicates
(donates, evictable, selected, clusterAdmin, shellScript, ...).
